not going to do this

+++ VirtualTourist/FlickrClient.swift
@@ -34,7 +34,7 @@ class FlickrClient {
       "has_geo=1",
    ]
    
-   func fetchPhotoPaths(pin: Pin, completionHandler: (paths: [String]?, numPages: Int?, errorString: String?) -> Void) {
+   func fetchPhotoPaths(pin: Pin, completionHandler: (paths: [String]?, numPages: Int32?, errorString: String?) -> Void) {
       
       parameters.append("lat=\(pin.latitude)")
       parameters.append("lon=\(pin.longitude)")
@@ -68,9 +68,9 @@ class FlickrClient {
          
          print("page=\(photosDict["page"]!) of \(photosDict["pages"]!)")
          
-         let numPages = photosDict["pages"] as! Int
+         let numPages = photosDict["pages"] as! Int32
          
-         pin.numPages = Int32(numPages)
+         pin.numPages = numPages
          
          // Generate flickr photo URLs from API JSON
          let paths = photoArray.map { (dict) -> String in
